requirements.txt
asyncio
requests
python-binance
mysql-connector-python